[REFACTOR] Move LoginRefresher into components folder
