+++ src/client/react/admin/AdminNavbar/SecondaryNavigation.js
@@ -9,7 +9,7 @@ import MenuButton from '../../../../../lib/react/components/MenuButton';
 import SettingsMenu from './SettingsMenu';
 import { withRouter } from 'react-router-dom';
 import { logout } from '../../../actions/authActions';
-import LoginRefresher from '../../sharedComponents/LoginRefresher';
+import LoginRefresher from '../../components/LoginRefresher';
 
 
 const mapStateToProps = (state, ownProps) => {
