+++ src/client/react/stages/components/HelpMenu.js
@@ -7,7 +7,7 @@ import { logout } from '../../../actions/authActions';
 import axios from 'axios';
 import API from '../../../API';
 import MenuButton from '../../../../../lib/react/components/MenuButton';
-import LoginRefresher from '../../sharedComponents/LoginRefresher';
+import LoginRefresher from '../../components/LoginRefresher';
 
 import '../../scss/components/_help-menu.scss';
